Move distribution.ts to /common
