+++ extensions/ql-vscode/src/common/distribution.ts
